refactor: match message with filters to message matcher
