+++ apps/api/src/app/events/usecases/index.ts
@@ -24,8 +24,10 @@ import { ShouldAddDigestJob } from './add-job/should-add-digest-job.usecase';
 import { SendMessageDelay } from './send-message/send-message-delay.usecase';
 import { SendTestEmail } from './send-message/test-send-email.usecase';
 import { MapTriggerRecipients, MapTriggerRecipientsCommand } from './map-trigger-recipients';
+import { MessageMatcher } from './trigger-event/message-matcher.service';
 
 export const USE_CASES = [
+  MessageMatcher,
   TriggerEvent,
   ProcessSubscriber,
   SendMessage,
